Move ValidatedStaticThisType helper type back into ResourceTypes
